Fix logic for calculating results
This commit will fix the issues with the behavior that did not let the
operands to be parsed correctly. I also reworked the logic for the
accumulator. Rather than having a separate variable for it, when the
user presses the equals sign, the result is stored on the first operand,
so that when other operations are performed, the next number will be
stored on the second operand and operated against the first.

I also fixed the logic for clearing out the screen when a new number is
inserted after pressing an operator.

What is left for me to implement is the logic for performing new
operations when the operator symbols are pressed. Then, I will have to
implement the logic for when the user attempts to divide by zero.