Rewrite logic for event handling I rewrote the event logic as it was messy, making it difficult to work out the logic for the numbers and operators.
I also made the input be an array. The idea is that
once a number is entered and an operator is pressed,
the number input is concatenated and converted to
a number that can be operated.

I still need to finish writing the logic for the
operators, but at least the operator function from
the previous implementation could be savaged. I
changed it slightly to make it cast the operands
to numbers, making it less messier. The event
handlers should handle the logic rather than letting
the functions handle it.

I may need to re-write this function later on to
separate concerns for the user inputs.

I also wrote some helper functions to clear the
input, set the screen to a value. I also modified
updateScreen() to simply cast the input array to
a string and display the result every time this
function is called.

The function for the point button and event handler
were also written.